handling table content horizontal overflow
