+++ src/constants/constants.ts
@@ -1,12 +1,1 @@
-export const COLUMN_DEFAULT_WIDTH: number = 150;
-
 export const DEFAULT_DELAY: number = 500;
-
-export const PAGE_SIZES: number[] = [5, 10, 25, 50, 100];
-
-export enum ColumnTypes {
-  String = "string",
-  Number = "number",
-  Boolian = "boolian",
-  Options = "options",
-}
