+++ src/features/columns/constants.ts
@@ -1,0 +1,10 @@
+export const COLUMN_DEFAULT_WIDTH: number = 150;
+
+export const COLUMNS_PADDING_X: number = 3; // ( units of 8px )
+
+export enum ColumnTypes {
+  String = "string",
+  Number = "number",
+  Boolian = "boolian",
+  Options = "options",
+}
