+++ src/features/data/constants.ts
@@ -1,0 +1,1 @@
+export const PAGE_SIZES: number[] = [5, 10, 25, 50, 100];
